refactor: 변경된 모달로 코드 변경
